+++ node.h
@@ -4,7 +4,7 @@
  * Data structure and procedure declarations for a weighted graph node.
  * 
  * Author: Richard Gale
- * Version: 27th August, 2022 
+ * Version: 28th August, 2022 
  */
 
 #ifndef NODE_H
@@ -13,6 +13,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdint.h>
+#include <inttypes.h>
 #include <stdbool.h>
 #include <math.h>
 
@@ -27,7 +28,7 @@ typedef struct node_data* node;
 /**
  * Initialises the node at the provided reference.
  */
-void node_init(node* n_ref, uint16_t x, uint16_t y, uint16_t z);
+void node_init(node* n_ref, uint8_t x, uint8_t y, uint8_t z);
 
 /**
  * Initialises the node's edges.
@@ -86,13 +87,13 @@ void node_set_came_from(node* n, node* came_from);
  * Sets the estimated total cost of a path if the path were to go 
  * through the node at the provided reference.
  */
-void node_set_f(node* n, uint32_t f);
+void node_set_f(node* n, uint64_t f);
 
 /**
  * Sets the cost of a path from the path's starting point to
  * the node at the provided reference.
  */
-void node_set_g(node* n, uint32_t g);
+void node_set_g(node* n, uint64_t g);
 
 /**
  * Adds a connection from one node to another, considering
